build(module): move store and effects module into module-games module
