+++ src/client/src/app/modules/games/module-games.module.ts
@@ -1,0 +1,26 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { AddGameComponent } from './components/add-game/add-game.component';
+import { RoutingModule } from './routing.module';
+import { gameFeatureKey, reducer } from 'src/app/store/reducers/game/game.reducer';
+import { EffectsModule } from '@ngrx/effects';
+import { StoreModule } from '@ngrx/store';
+import { GameEffects } from 'src/app/store/effects/game/game.effects';
+
+
+
+@NgModule({
+  declarations: [
+    AddGameComponent,
+  ],
+  imports: [
+    CommonModule,
+    FormsModule,
+    ReactiveFormsModule,
+    RoutingModule,
+    EffectsModule.forFeature([GameEffects]),
+    StoreModule.forFeature(gameFeatureKey, reducer),
+  ]
+})
+export class GamesModule { }
