+++ src/client/src/app/modules/games/routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule, Routes } from '@angular/router';
-import { AddGameComponent } from '../components/add-game/add-game.component';
+import { AddGameComponent } from './components/add-game/add-game.component';
 import { PageGamesComponent } from 'src/app/pages/page-games/page-games.component';
 
 const routes: Routes = [
